requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
drf-yasg==1.21.14
psycopg2-binary==2.9.11
gunicorn==22.0.0
whitenoise==6.7.0
